可seek，但连续点击seek时画面卡死

+++ app/src/main/java/com/skystack/skymediaplayer/MediaPlayer/IMediaPlayer.java
@@ -26,7 +26,7 @@ public interface IMediaPlayer {
 
     boolean isPlaying();
 
-    void seekTo(long msec) throws IllegalStateException;
+    void seekTo(int msec) throws IllegalStateException;
 
     int getCurrentPosition();
